Defined show method to display the team's name, manager's, coach's, lineup and bench information.
Changed lineup and bench to be dynamically allocated incase a nullptr is given in constructor, and then assigned nullptr to each slot in the newly created arrays.

if a coach/manager are given instead of nullptr, sets their team as the team being created in this constructor.

Changed the destructor to delete benchplayers and lineup as they're now dynamically allocated.

defined the "addPlayer" function which adds players to lineup if it's not full, otherwise to bench (if its not full).