Changed .show() method to use vector iterator to show each player instead of the for loop.
Changed constructor, removed the phys size variables and moved the vector lineup and vector benchplayers before other variables that have default input.
Removed the condition that if input of lineup and bench is null to not dynamically allocate (as vector makes it on its own)

Changed addPlayer

Changed isReady to check vector phys size.